+++ packages/tests/specs/InteractionsMaterial.test.js
@@ -1,30 +1,27 @@
 import * as configs from "../support/configs";
 import * as inits from "../support/inits";
-import { with_qb_material, sleep } from "../support/utils";
-import Autocomplete from "@material-ui/lab/Autocomplete";
+import { with_qb_material } from "../support/utils";
 import { expect } from "chai";
-
-const stringifyOptions = (options) => {
-  return options.map(({title, value}) => `${value}_${title}`).join(";");
-};
+import { autocompleteTestsFor } from "./Autocomplete";
 
 describe("interactions on Material-UI", () => {
 
   describe("autocomplete", () => {
-    it("find B", async () => {
-      await with_qb_material(configs.with_autocomplete, inits.with_autocomplete_a, "JsonLogic", async (qb, onChange, {expect_jlogic}) => {
-        let ac = qb.find(Autocomplete).filter({label: "Select value"});
-        expect(stringifyOptions(ac.prop("options"))).to.eq("a_a");
-        
-        ac.prop("onInputChange")(null, "b");
-        await sleep(200); // should be > 50ms delay
-        qb.update();
-        ac = qb.find(Autocomplete).filter({label: "Select value"});
+    const {
+      testsSingleStrict,
+      testsMultipleStrict,
+    } = autocompleteTestsFor("mui", 4, it, with_qb_material);
 
-        expect(stringifyOptions(ac.prop("options"))).to.eq("a_a;b_B");
-      });
+    describe("single-strict", () => {
+      testsSingleStrict();
+    });
+
+    describe("multiple-strict", () => {
+      testsMultipleStrict();
     });
   });
+
+  //-------
 
   it("should render labels with showLabels=true", async () => {
     await with_qb_material([configs.with_different_groups, configs.with_settings_show_labels], inits.with_different_groups, "JsonLogic", (qb) => {