+++ CHANGELOG.md
@@ -11,6 +11,13 @@
     - Added new test app `sandbox_next` to demonstrate new server-side features
   - Export utils (like `mongoFormatOp1`, `mongoFormatOp2`) in `CoreConfig.ctx.utils` (PR #866) (issue #890)
   - 2nd parameter of `onChange` callback now equals original config, not extended config (PR #866) (issue #364)
+  - Fix totalQuery structure in mongodb (PR #916) (issue #915)
+  - Fixed bug in `loadTree()` (PR #917) (issue #356)
+  - Fixed bug on change group value (PR #944) (issue #923)
+  - Fixed render of option titles in Autocomplete AntDesign (PR #947) (issues #930, #942)
+  - Implemented `allowCustomValues` for MUI (PR #947) (issue #327)
+  - Improved autocomplete for MUI and AntD. Mark custom values with coral color (PR #947)
+  - Type fixes for `strictNullChecks` (PR #948) (issue #945)
 - 6.2.0
   - Fixed type `Config`: should have render settings like `renderSize` (PR #909) (issue #879)
   - Fixed type for `renderBeforeWidget`: `RuleProps` instead of wrong `FieldProps` (PR #909) (issue #879)